# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Adare General Hospital Platform API",
    "version": "1.0.0",
    "description": "REST API for the Adare General Hospital digital platform (Hawassa, Sidama Region, Ethiopia). All responses use the envelope { ok, data, message } or { ok:false, code, error }. Protected endpoints require Authorization: Bearer <access_token>; refresh tokens rotate via HTTP-only cookie on /api/auth."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
...
  },
  "paths": {
...
    "/news": { "get": { "summary": "Published news (paginated)", "parameters": [{ "name": "page", "in": "query" }, { "name": "category", "in": "query" }, { "name": "q", "in": "query" }], "responses": { "200": { "description": "Articles" } } } },
    "/news/{slug}": { "get": { "summary": "Single article", "parameters": [{ "name": "slug", "in": "path", "required": true }], "responses": { "200": { "description": "Article" }, "404": { "description": "Not found" } } } },
...
    "/admin/dashboard": { "get": { "security": [{ "bearerAuth": [] }], "summary": "KPIs + chart data", "responses": { "200": { "description": "Dashboard" } } } },
    "/admin/notifications": { "get": { "security": [{ "bearerAuth": [] }], "summary": "Staff notification feed", "responses": { "200": { "description": "Feed + unread count" } } } },
    "/admin/users": {
      "get": { "security": [{ "bearerAuth": [] }], "summary": "List staff (admin)", "responses": { "200": { "description": "Users" } } },
      "post": { "security": [{ "bearerAuth": [] }], "summary": "Create staff user (admin)", "responses": { "201": { "description": "User" } } }
    },
    "/admin/news": { "get": { "security": [{ "bearerAuth": [] }], "summary": "CMS: list articles" }, "post": { "security": [{ "bearerAuth": [] }], "summary": "CMS: create article" } },
    "/admin/audit": { "get": { "security": [{ "bearerAuth": [] }], "summary": "Audit log (admin only)", "responses": { "200": { "description": "Log entries" } } } },
...
  }
}